refactor alert to single component
