+++ src/app/app.component.html
@@ -2,6 +2,6 @@
   <app-nav></app-nav>
   <div class="content">
     <router-outlet></router-outlet>
-    <app-snack-bar></app-snack-bar>
+    <app-alert></app-alert>
   </div>
 </div>
