--- src/app/shared/components/alert.component.html
@@ -1,4 +1,0 @@
-<div>
-  <h2>{{ title() }}</h2>
-  <div class="mat-subtitle-2" *ngFor="let line of message()">{{ line }}</div>
-</div>
