--- src/app/shared/components/app-snack-bar.component.html
@@ -1,1 +1,0 @@
-<div></div>
